desktop/requirements.txt
-r ../backend/requirements.txt
pywebview==5.3.2
pyinstaller==6.11.1
